Implement full CRUD for Auth (users)

+++ server/src/services/authService.js
@@ -1,7 +1,8 @@
 const bcrypt = require('bcrypt');
-// ====================================================
-const { User } = require('../db/models');
-// ====================================================
+// ==============================================================
+const { User, Token } = require('../db/models');
+// ==============================================================
+const { SALT_ROUNDS } = require('../constants');
 const {
   generateTokens,
   saveToken,
@@ -11,15 +12,26 @@ const {
 } = require('./tokenService');
 const { badRequest, unAuthorizedError } = require('../errors/authError');
 
+async function hashPassword(password) {
+  return await bcrypt.hash(password, SALT_ROUNDS);
+}
+
+function emailToLowerCase(email) {
+  return email.toLowerCase();
+}
+
 class AuthService {
   async registration(fullName, email, password) {
-    const person = await User.findOne({ where: { email } });
+    const emailToLower = emailToLowerCase(email);
+
+    const person = await User.findOne({ where: { email: emailToLower } });
     if (person) throw badRequest('This user already exists');
 
-    const hashedPassword = await bcrypt.hash(password, 10); // Хешируем пароль
+    const hashedPassword = await hashPassword(password);
+
     const user = await User.create({
       fullName,
-      email,
+      email: emailToLower,
       password: hashedPassword,
     });
 
@@ -27,13 +39,14 @@ class AuthService {
       user: {
         id: user.id,
         email: user.email,
-        roleId: user.roleId,
       },
     };
   }
 
   async login(email, password) {
-    const user = await User.findOne({ where: { email } });
+    const emailToLower = emailToLowerCase(email);
+
+    const user = await User.findOne({ where: { email: emailToLower } });
     if (!user) throw unAuthorizedError();
 
     const isPassRight = await bcrypt.compare(password, user.password);
@@ -48,13 +61,16 @@ class AuthService {
       ...tokens,
       user: {
         id: userId,
-        email,
+        email: emailToLower,
       },
     };
   }
 
   async logout(refreshToken) {
+    if (!refreshToken) throw unAuthorizedError();
+
     const token = await deleteToken(refreshToken);
+
     return token;
   }
 
@@ -67,10 +83,12 @@ class AuthService {
     if (!data || !dbToken) throw unAuthorizedError();
 
     const { email } = data;
-    const user = await User.findOne({ where: { email } });
+    const emailToLower = emailToLowerCase(email);
+
+    const user = await User.findOne({ where: { email: emailToLower } });
     const userId = user.id;
 
-    const tokens = generateTokens({ email });
+    const tokens = generateTokens({ email: emailToLower });
 
     await saveToken(userId, tokens.refreshToken);
 
@@ -78,25 +96,96 @@ class AuthService {
       ...tokens,
       user: {
         id: userId,
-        email,
+        email: emailToLower,
       },
     };
   }
 
   async getAllUsers() {
     const users = await User.findAll();
+
+    if (users.length === 0) {
+      throw badRequest('Users not found');
+    }
+
     return users;
   }
 
-  async deleteUser(email) {
-    const user = await User.findOne({ where: { email } });
+  async getUserById(id) {
+    const user = await User.findByPk(id);
 
-    if (user) {
-      const delUser = await User.destroy({ where: { email } });
-      return delUser;
+    if (!user) {
+      throw badRequest('User not found');
     }
 
-    return null;
+    return user;
+  }
+
+  async updateUser(id, fullName, email, password) {
+    const user = await User.findOne({ where: { id } });
+
+    if (!user) {
+      throw badRequest('User not found');
+    }
+
+    const updateData = { fullName };
+
+    if (email && email.toLowerCase() !== user.email.toLowerCase()) {
+      const newEmail = emailToLowerCase(email);
+
+      const person = await User.findOne({ where: { email: newEmail } });
+      if (person) throw badRequest('This email is already used');
+
+      updateData.email = newEmail;
+
+      const token = await Token.findOne({ where: { userId: id } });
+
+      await deleteToken(token.refreshToken);
+    }
+
+    if (password) {
+      const hashedPassword = await hashPassword(password);
+      updateData.password = hashedPassword;
+    }
+
+    const [affectedRows, [updatedUsers]] = await User.update(updateData, {
+      where: { id },
+      returning: true,
+    });
+
+    if (affectedRows === 0) {
+      throw badRequest('User not found');
+    }
+
+    return {
+      user: {
+        id: updatedUsers.id,
+        email: updatedUsers.email,
+      },
+    };
+  }
+
+  async deleteUser(id) {
+    const user = await User.findByPk(id);
+
+    if (!user) {
+      throw badRequest('User not found');
+    }
+
+    const refreshTokenRecord = await Token.findOne({
+      where: { userId: user.id },
+    });
+
+    const refreshToken = refreshTokenRecord
+      ? refreshTokenRecord.refreshToken
+      : null;
+
+    if (refreshToken) {
+      await deleteToken(refreshToken);
+    }
+
+    const delUser = await User.destroy({ where: { id } });
+    return delUser;
   }
 }
